Update: Product Domain 에서 productStatus 제거
